\name{FGSG-package}
\alias{FGSG}
\docType{package}
\title{
Feature grouping and selection over an undirected graph
}
\description{
FGSG package implements algorithms for feature grouping and selection over an undirected graph. This package can work under Linux enviornment, but is not guaranteed under Windows.
}
\details{
\tabular{ll}{
Package: \tab FGSG\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-10\cr
License: \tab GPL-2\cr
}

}
\author{
Xiaotong Shen, Yiwen Sun

Maintainer: Yiwen Sun <sunxx847@umn.edu>
}
\references{
[1]S.Kim and E.Xing. Statisctical estimation of correlated genome associations to a quantitative trait network. PLoS genetics, 5(8):e1000587, 2009

[2]S.Yang, L.Yuan, Y.Lai, X.Shen, P.Wonka, and J.Ye. Feature grouping and selection over an undirected graph. KDD, 2012.
}

